+++ src/App.jsx
@@ -1,4 +1,5 @@
 import { store } from "@/Store";
+import { useEffect } from "react";
 import { Provider } from "react-redux";
 import {
   Ads,
@@ -10,10 +11,12 @@ import {
   Testimonials,
   Benefits,
 } from "./Components";
+import ProyectRoutes from "./ProyectRoutes";
 
 function App() {
   return (
     <>
+      <ProyectRoutes />
       <Provider store={store}>
         <Layout>
           <Banner />
