+++ src/Components/Layout/Layout.jsx
@@ -1,16 +1,16 @@
-import { Box } from '@chakra-ui/react'
-import { Footer, Header, Navbar } from '..'
+import { Box } from "@chakra-ui/react";
+import { Footer, Header, Navbar } from "..";
 
 function Layout({ children }) {
   return (
     <>
       <Header />
       <Navbar />
-      <Box maxW={'1440px'} mx={'auto'}>
+      <Box maxW={"1440px"} mx={"auto"}>
         {children}
       </Box>
       <Footer />
     </>
-  )
+  );
 }
-export default Layout
+export default Layout;
