+++ src/Components/NewProducts/NewProducts.css
@@ -1,0 +1,8 @@
+.slick-prev:before {
+  color: rgb(148, 148, 148);
+  font-size: 20px;
+}
+.slick-next:before {
+  color: rgb(148, 148, 148);
+  font-size: 20px;
+}
