+++ src/Components/index.js
@@ -2,6 +2,7 @@ export { default as Ads } from "./Ads/Ads";
 export { default as Banner } from "./Banner/Banner";
 export { default as ImageSlider } from "./Banner/ImageSlider";
 export * from "./Banner/SlideData";
+export { default as Benefits } from "./Benefits/Benefits";
 export { default as FollowUs } from "./FollowUs/FollowUs";
 export { default as AccordionFooter } from "./Layout/Footer/AccordionFooter";
 export { default as Footer } from "./Layout/Footer/Footer";
@@ -15,4 +16,3 @@ export { default as NewProducts } from "./NewProducts/NewProducts";
 export { default as StarRating } from "./NewProducts/StarRating";
 export { default as Sponsors } from "./Sponsors/Sponsors";
 export { default as Testimonials } from "./Testimonials/Testimonials";
-export { default as Benefits } from "./Benefits/Benefits";
