+++ src/Helpers/pagesRoutes.js
@@ -1,0 +1,12 @@
+const pagesRoutes = [
+  "/Laptops",
+  "/DesktopPC",
+  "/NetworkinDevices",
+  "/Printer&Scanners",
+  "/PcParts",
+  "/AllOther Products",
+  "/Repairs",
+  "/OurDeals",
+];
+
+export default pagesRoutes;
