+++ src/Pages/Laptops.jsx
@@ -3,7 +3,7 @@ import { Box, Text } from "@chakra-ui/react";
 function Laptops() {
   return (
     <Box>
-      <Text style={{ backgroundColor: "red" }}>Laptops</Text>
+      <Text>Laptops</Text>
     </Box>
   );
 }
